+++ test/test_xml.sh
@@ -43,5 +43,11 @@ EOF
 new "xml optional encode single and double quote"
 expecteof "$PROG" 0 "$XML" "^<message>To allow attribute values to contain both single and double quotes, the apostrophe or single-quote character ' may be represented as ' and the double-quote character as \"</message>$"
 
+new "Double quotes for attributes"
+expecteof "$PROG" 0 '<x a="t"/>' '^<x a="t"/>$'
+
+new "Single quotes for attributes (returns double quotes but at least parses right)"
+expecteof "$PROG" 0 "<x a='t'/>" '^<x a="t"/>$'
+
 rm -rf $dir